--- mlrun/config/__init__.py
@@ -1,27 +1,0 @@
-# Copyright 2018 Iguazio
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-Configuration system.
-
-Configuration can be in either a configuration file specified by
-MLRUN_CONFIG_FILE environment variable or by environment variables.
-
-Environment variables are in the format "MLRUN_httpdb__port=8080". This will be
-mapped to config.httpdb.port. Values should be in JSON format.
-"""
-
-from .default import default_config  # noqa
-from .dynamic import config  # noqa
-from .loader import env_file_key, env_prefix  # noqa
-from .static import static_config  # noqa
